selenium_ui_tests_bdd/requirements.txt
selenium==3.141.0
pytest==4.3.1
allure-pytest==2.6.1
allure-python-commons==2.6.1
pytest-bdd==3.1.0